# docs/openapi.yaml
# Generated by .github/workflows/api-docs.yml from the springdoc
# endpoint /v3/api-docs.yaml. Do not edit by hand — change the
# annotations in src/main/java/**/infrastructure/web/ instead.
openapi: 3.1.0
info:
  title: Presence System API
  description: |
    REST API of the **Attendance Management and Verification System** — an NFC-based attendance tracker built with Java 21 / Spring Boot, following a simplified hexagonal architecture.

    The API exposes two families of endpoints:

    * `/api/v1/attendances` — **public**, called by the Python NFC client running on the ACR122U reader. One call per badge scan.
    * `/api/admin/**` — **protected**, requires HTTP Basic authentication with the `ADMIN` role. Used by the web administration dashboard for manual CRUD on attendance records.

    ### Authentication

    Admin endpoints use **HTTP Basic**. Send the `Authorization: Basic <base64(user:password)>` header on every request. CSRF is disabled and sessions are created on demand.

    ### CORS

    Only `http://localhost:*` and `http://127.0.0.1:*` origins are allowed, with credentials.
  contact:
    name: Eric Beaubrun
    url: https://github.com/ericbeaubrun/presence-system
  license:
    name: See repository
    url: https://github.com/ericbeaubrun/presence-system
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Local Docker Compose stack
security:
- basicAuth: []
tags:
- name: Attendance
  description: Public endpoint used by the NFC reader client.
- name: Admin
  description: "Administration endpoints, `ADMIN` role required."
paths:
  /api/admin/attendances:
    get:
      tags:
      - Admin
      summary: List every attendance record
      description: "Returns the full attendance history, enriched with student, course\
        \ and room details."
      operationId: getAllAttendances
      responses:
        "401":
          description: Missing or invalid credentials.
        "403":
          description: "Authenticated, but the user is not an `ADMIN`."
        "200":
          description: Attendance list.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceDto"
    put:
      tags:
      - Admin
      summary: Update an existing attendance record
      description: |
        The record is identified by the `studentId` + `courseId` pair carried in the body. Status, arrival time and justification are overwritten.
      operationId: updateAttendance
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceUpdateRequest"
        required: true
      responses:
        "200":
          description: Record updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceDto"
        "404":
          description: No record matches the given `studentId` / `courseId`.
        "401":
          description: Missing or invalid credentials.
        "403":
          description: "Authenticated, but the user is not an `ADMIN`."
    post:
      tags:
      - Admin
      summary: Create an attendance record manually
      operationId: createAttendance
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceUpdateRequest"
        required: true
      responses:
        "201":
          description: Record created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceDto"
        "400":
          description: "Validation failed, or the student / course does not exist."
        "401":
          description: Missing or invalid credentials.
        "403":
          description: "Authenticated, but the user is not an `ADMIN`."
    delete:
      tags:
      - Admin
      summary: Delete an attendance record
      operationId: deleteAttendance
      parameters:
      - name: studentId
        in: query
        description: Student identifier.
        required: true
        schema:
          type: string
        example: E001
      - name: courseId
        in: query
        description: Course identifier.
        required: true
        schema:
          type: string
        example: C001
      responses:
        "200":
          description: Record deleted.
          content:
            application/json:
              example:
                message: Pointage supprimé avec succès.
        "404":
          description: No record matches the given `studentId` / `courseId`.
          content:
            application/json: {}
        "401":
          description: Missing or invalid credentials.
        "403":
          description: "Authenticated, but the user is not an `ADMIN`."
  /api/v1/attendances:
    post:
      tags:
      - Attendance
      summary: Register an attendance from an NFC card scan
      description: |
        Called by the hardware client when a card is presented to the reader.

        The server resolves the student from the card UID, derives the room from the reader id (`L` prefix is replaced by `S`), looks up the course currently running in that room, and records a `present` attendance with the current time.

        The request is rejected when the card is unknown, when no course is scheduled in that room at that moment, or when the student already checked in for that course.
      operationId: scanCard
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceRequest"
        required: true
      responses:
        "200":
          description: Attendance registered.
          content:
            text/plain:
              schema:
                type: string
              example: Présence enregistrée avec succès pour l'étudiant E001 au cours
                C001
        "400":
          description: |
            Business rule violation or validation error. The body carries the reason: unknown card, no course scheduled in the room right now, or duplicate check-in.
          content:
            text/plain:
              schema:
                type: string
              examples:
                unknownCard:
                  description: unknownCard
                  value: "Aucun étudiant trouvé avec la carte NFC : 04A3B2C1"
                noCourse:
                  description: noCourse
                  value: Aucun cours n'est programmé dans cette salle en ce moment.
                duplicate:
                  description: duplicate
                  value: L'étudiant a déjà été enregistré (présent) pour ce cours.
        "500":
          description: Unexpected server error.
          content:
            text/plain:
              schema:
                type: string
              example: "Erreur serveur : ..."
  /api/admin/me:
    get:
      tags:
      - Admin
      summary: Return the authenticated administrator
      description: |
        Validates the supplied credentials and echoes back the identity of the caller. The dashboard uses it as a login check.
      operationId: me
      responses:
        "401":
          description: Missing or invalid credentials.
        "403":
          description: "Authenticated, but the user is not an `ADMIN`."
        "200":
          description: Credentials are valid.
          content:
            application/json:
              example:
                username: admin
                role: ROLE_ADMIN
components:
  schemas:
    AttendanceUpdateRequest:
      type: object
      description: Payload used by the dashboard to create or update a record.
      properties:
        studentId:
          type: string
          example: E001
          minLength: 1
        courseId:
          type: string
          example: C001
          minLength: 1
        status:
          type: string
          description: Only these two values are accepted.
          enum:
          - present
          - absent justifie
          example: present
          minLength: 1
          pattern: present|absent justifie
        arrivalTime:
          type:
          - string
          - "null"
          description: Arrival time as `HH:mm:ss`. Nullable.
          example: 08:15:00
        justification:
          type:
          - string
          - "null"
          description: Free-text justification. Nullable.
          example: Certificat médical
      required:
      - courseId
      - status
      - studentId
    AttendanceDto:
      type: object
      description: "Attendance record enriched with student, course and room details."
      properties:
        studentId:
          type: string
          example: E001
        studentFullName:
          type: string
          example: Eric Beaubrun
        courseId:
          type: string
          example: C001
        courseDate:
          type: string
          example: 2026-08-05
        courseStartTime:
          type: string
          example: 08:00:00
        courseEndTime:
          type: string
          example: 10:00:00
        roomId:
          type: string
          example: S1101
        status:
          type: string
          example: present
        arrivalTime:
          type:
          - string
          - "null"
          example: 08:15:00
        justification:
          type:
          - string
          - "null"
    AttendanceRequest:
      type: object
      description: Payload emitted by the NFC reader client on each scan.
      properties:
        readerId:
          type: string
          description: "Reader identifier, exactly 5 characters. The room id is derived\
            \ from it."
          example: L1101
          maxLength: 5
          minLength: 5
        cardId:
          type: string
          description: "NFC card UID, exactly 8 characters."
          example: 04A3B2C1
          maxLength: 8
          minLength: 8
      required:
      - cardId
      - readerId
  securitySchemes:
    basicAuth:
      type: http
      description: HTTP Basic credentials of a user holding the `ADMIN` role.
      scheme: basic